add the DA id to the incoming arguments of the MERGE_START ioctl (trac-4377)
As it stands, the structure passed to the MERGE_START ioctl does not
include the DA id, it only includes an array of the CT ids. While this
is enough to guarantee correct operation, it makes error reporting
less clear: if one or more of the CTs given do not exist, is it
because they themselves have disappeared in the meantime (or have
never existed), or is it because the entire DA which they were
supposed to belong to has ceased to exist? Thus the purpose of this
change is to make it easier to report these two as separate error
conditions.

The way this is achieved is of course by also passing the DA id when
requesting the start of a new merge.
(transplanted from 4ca0dc2a7e0cc9f4c38bcdfd990a052243df68e8)

--HG--
branch : v2
extra : transplant_source : L%A0%DC%2A%7E%0C%C9%F4%C3%8B%CD%FD%99%0A%05%22C%DFh%E8